GatewayIoT/requirements.txt
fastapi
uvicorn
python-socketio
pyserial
paho-mqtt>=2.0.0
python-dotenv
aiofiles
